@@ -1,21 +1,20 @@
 module github.com/junegunn/fzf
 
 require (
-	github.com/gdamore/tcell/v2 v2.5.4
-	github.com/mattn/go-isatty v0.0.17
-	github.com/mattn/go-runewidth v0.0.14
-	github.com/mattn/go-shellwords v1.0.12
-	github.com/rivo/uniseg v0.4.2
-	github.com/saracen/walker v0.1.3
-	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab
-	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211
+	github.com/charlievieth/fastwalk v1.0.9
+	github.com/gdamore/tcell/v2 v2.8.1
+	github.com/junegunn/go-shellwords v0.0.0-20250127100254-2aa3b3277741
+	github.com/mattn/go-isatty v0.0.20
+	github.com/rivo/uniseg v0.4.7
+	golang.org/x/sys v0.30.0
+	golang.org/x/term v0.29.0
 )
 
 require (
-	github.com/gdamore/encoding v1.0.0 // indirect
+	github.com/gdamore/encoding v1.0.1 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
-	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4 // indirect
-	golang.org/x/text v0.5.0 // indirect
+	github.com/mattn/go-runewidth v0.0.16 // indirect
+	golang.org/x/text v0.21.0 // indirect
 )
 
-go 1.17
+go 1.20
